refactor(diagnostics): extract runtime descriptors to diagnostics.info
Move DiagnosticInfo and DiagnosticParameterInfo out of diagnostics.metadata
into a new diagnostics.info package. They are runtime descriptors that depend
on configuration (i18n, LSP mapping) and BSLDiagnostic, unlike the metadata
package which holds the pure, dependency-free vocabulary (the @DiagnosticMetadata
annotation and enums).

This makes diagnostics.metadata a clean leaf: configuration can depend on the
metadata vocabulary (for deserializing diagnostic-metadata overrides) without
pulling configuration-coupled runtime code, removing the back edge that bound
metadata into the configuration cycle.

Note: docs generation (generateDiagnosticDocs) requires bslls-dev-tools that
resolves DiagnosticInfo/DiagnosticParameterInfo from diagnostics.info (with a
fallback to diagnostics.metadata). Until that plugin release is wired in, the
generateDiagnosticDocs task fails; compilation, tests and spotless pass.
