+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/diagnostics/BSLDiagnostic.java
@@ -22,7 +22,7 @@
 package com.github._1c_syntax.bsl.languageserver.diagnostics;
 
 import com.github._1c_syntax.bsl.languageserver.context.DocumentContext;
-import com.github._1c_syntax.bsl.languageserver.diagnostics.metadata.DiagnosticInfo;
+import com.github._1c_syntax.bsl.languageserver.diagnostics.info.DiagnosticInfo;
 import org.eclipse.lsp4j.Diagnostic;
 
 import java.util.List;
