+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/diagnostics/info/DiagnosticParameterInfo.java
@@ -19,8 +19,9 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with BSL Language Server.
  */
-package com.github._1c_syntax.bsl.languageserver.diagnostics.metadata;
+package com.github._1c_syntax.bsl.languageserver.diagnostics.info;
 
+import com.github._1c_syntax.bsl.languageserver.diagnostics.metadata.DiagnosticParameter;
 import lombok.Getter;
 
 import java.lang.reflect.Field;
