+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/reporters/CodeQualityReportEntry.java
@@ -22,7 +22,7 @@
 package com.github._1c_syntax.bsl.languageserver.reporters;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.github._1c_syntax.bsl.languageserver.diagnostics.metadata.DiagnosticInfo;
+import com.github._1c_syntax.bsl.languageserver.diagnostics.info.DiagnosticInfo;
 import com.github._1c_syntax.bsl.languageserver.diagnostics.metadata.DiagnosticSeverity;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
